+++ NEWS.md
@@ -7,6 +7,8 @@
 - Better documentation of `argonDashPage()`
 
 ## Minor improvement
+- add `argonTable()`, `argonTableItems` and `argonTableItem`
+- add `argonAvatars()` and `argonAvatar` 
 - Fix #12: reduce vertical space between brand logo and menu items in `argonDashSidebar()`. Thanks @ShixiangWang
 - fix issue in `argonInfoCard()` when stat is NULL
 - fix #3 thanks @noahbishop: when vertical is FALSE in the `argonDashSidebar()`,